requirements.txt
# Python dependencies for Personal Budget Tracker
# Backend
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=24.3
django-cors-headers>=4.4

# Utility (used to extract the assessment PDF)
pypdf>=4.0.0

